# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for damarmunda, a camunda golang",
        "title": "damarmunda API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/bpmn/pembuatan_media_berita_technology": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "Search PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit 20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page 1",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort desc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Since desc",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Until desc",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology Search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "Create PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "description": "PembuatanMediaBeritaTechnology Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreatePembuatanMediaBeritaTechnologyPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PembuatanMediaBeritaTechnology Created",
                        "schema": {
                            "$ref": "#/definitions/main.DataStorePembuatanMediaBeritaTechnologyWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/bpmn/pembuatan_media_berita_technology/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetById PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "GetById PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID from table",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology GetById",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "Cancel PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ProcessInstanceKey",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology Canceled",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "Update PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID from table",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PembuatanMediaBeritaTechnology Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdatePembuatanMediaBeritaTechnologyPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology Updated",
                        "schema": {
                            "$ref": "#/definitions/main.DataStorePembuatanMediaBeritaTechnologyWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/bpmn/pembuatan_media_berita_technology/{id}/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetHistoryById PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "GetHistoryById PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID from table",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Size 50",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order DESC ASC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type USER_TASK",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "State ACTIVE",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort startDate",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchAfter 1731486859777,2251799814109407",
                        "name": "searchAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchBefore 1731486859777,2251799814109407",
                        "name": "searchBefore",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology GetHistoryById",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/bpmn/pembuatan_media_berita_technology/{id}/incidents": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetProcessIncidents PembuatanMediaBeritaTechnology",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bpmn/PembuatanMediaBeritaTechnology"
                ],
                "summary": "GetProcessIncidents PembuatanMediaBeritaTechnology",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID from table",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Size 50",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order DESC ASC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type USER_TASK",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "State ACTIVE",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort startDate",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchAfter 1731486859777,2251799814109407",
                        "name": "searchAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchBefore 1731486859777,2251799814109407",
                        "name": "searchBefore",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PembuatanMediaBeritaTechnology GetProcessIncidents",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.CreatePembuatanMediaBeritaTechnologyPayload": {
            "type": "object",
            "properties": {
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "main.DataStorePembuatanMediaBeritaTechnologyWrapper": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/store.PembuatanMediaBeritaTechnology"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "main.UpdatePembuatanMediaBeritaTechnologyPayload": {
            "type": "object",
            "properties": {
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}